polymorpher::measurer

Trait Measurer

pub trait Measurer {
    // Required methods
    fn measure_cubic(&self, c: &Cubic) -> f32;
    fn find_cubic_cut_point(&self, c: &Cubic, m: f32) -> f32;
}

Required Methods

fn measure_cubic(&self, c: &Cubic) -> f32

Returns size of given cubic, according to however the implementation wants to measure the size (angle, length, etc). It has to be greater or equal to 0.

fn find_cubic_cut_point(&self, c: &Cubic, m: f32) -> f32

Given a cubic and a measure that should be between 0 and the value returned by measureCubic (If not, it will be capped), finds the parameter t of the cubic at which that measure is reached.