polymorpher

Module polygon_builder

Structs

Circle

Pill

PillStar

Rectangle

RoundedPolygonBuilder

Provides a more convenient and familiar way to create certain shapes using builder pattern, helping to avoid functions with hellish Option<T> arguments.

Star

Traits

HaveRadius

HaveRounding

HaveSize

Functions

pill_star_vertices_from_num_verts 🔒

star_vertices_from_num_verts 🔒