polymorpher

Struct RoundedPolygonBuilder

pub struct RoundedPolygonBuilder<T> {
    pub(crate) data: T,
    pub(crate) center: Point,
    pub(crate) rounding: CornerRounding,
    pub(crate) per_vertex_rounding: Vec<CornerRounding>,
}

Provides a more convenient and familiar way to create certain shapes using builder pattern, helping to avoid functions with hellish Option<T> arguments.

Currently supports the creation of the following shapes:

Rectangles

Circles

Stars

Pills

Pill stars

Fields
data: Tcenter: Pointrounding: CornerRoundingper_vertex_rounding: Vec<CornerRounding>
Implementations

impl<T> RoundedPolygonBuilder<T>

pub const fn with_center(self, center: Point) -> Self

Sets the center point for the resulting rounded polygon.

impl<T: HaveRounding> RoundedPolygonBuilder<T>

pub const fn with_rounding(self, rounding: CornerRounding) -> Self

Sets the common CornerRounding for each of the vertices.

pub fn with_rounding_per_vertex<I: IntoIterator<Item = CornerRounding>>( self, iter: I, ) -> Self

Sets CornerRounding for each individual vertex. The size of the iterator must be equal to the number of vertices.

impl<T: HaveSize> RoundedPolygonBuilder<T>

pub fn with_size(self, size: Size) -> Self

Sets the size of the resulting shape.

pub fn with_width(self, width: f32) -> Self

Sets the width of the resulting shape.

pub fn with_height(self, height: f32) -> Self

Sets the height of the resulting shape.

impl<T: HaveRadius> RoundedPolygonBuilder<T>

pub fn with_radius(self, radius: f32) -> Self

Sets the radius of the resulting shape.

impl RoundedPolygonBuilder<Circle>

pub const fn with_vertices(self, vertices: usize) -> Self

Sets the number of vertices in the resulting polygon with which to approximate the circle.

pub fn build(self) -> RoundedPolygon

Consuming itself, it creates and returns RoundedPolygon, which forms a circle.

impl RoundedPolygonBuilder<Rectangle>

pub const fn with_rounding(self, rounding: CornerRounding) -> Self

Sets the common CornerRounding for each of the rectangle corners.

pub fn with_rounding_per_vertex(self, corners: [CornerRounding; 4]) -> Self

Sets CornerRounding for each individual corner. The size of the iterator must be equal to 4 for four rectangle corners.

pub fn build(self) -> RoundedPolygon

Consuming itself, it creates and returns RoundedPolygon, which forms a rectangle.

impl RoundedPolygonBuilder<Star>

pub const fn with_inner_radius(self, radius: f32) -> Self

Set inner radius for the resulting star shape, which must be greater than 0 and less than or equal to radius.

pub const fn with_inner_rounding(self, rounding: CornerRounding) -> Self

Sets the common CornerRounding for each of the vertices on the inner radius.

pub fn build(self) -> RoundedPolygon

Consuming itself, it creates and returns RoundedPolygon, which forms a star.

impl RoundedPolygonBuilder<Pill>

pub const fn with_smoothing(self, smoothing: f32) -> Self

Sets the amount by which the arc is “smoothed” by extending the curve from the circular arc on each endcap to the edge between the endcaps.

pub fn build(self) -> RoundedPolygon

Consuming itself, it creates and returns RoundedPolygon, which forms a pill.

impl RoundedPolygonBuilder<PillStar>

pub const fn with_vertices_per_radius(self, count: usize) -> Self

pub const fn with_vertex_spacing(self, spacing: f32) -> Self

Sets the vertex spacing factor, which determines how vertices on the circular ends are laid out along the outline.

pub const fn with_start_location(self, location: f32) -> Self

Sets the start location which determines how far along the perimeter of the resulting shape to start the underlying curves of which it is comprised.

pub const fn with_inner_radius_ratio(self, radius: f32) -> Self

Sets the inner radius ratio for the resulting star shape, must be greater than 0 and less than or equal to 1.

pub const fn with_inner_rounding(self, rounding: CornerRounding) -> Self

Sets the common CornerRounding for each of the vertices on the inner radius ratio.

pub fn build(self) -> RoundedPolygon

Consuming itself, it creates and returns RoundedPolygon, which forms a pill star.